CAMBIANDO A SQL

+++ Super/src/julenmartin/accesodatos/Dao.java
@@ -3,11 +3,11 @@ package julenmartin.accesodatos;
 public interface Dao<T> {
 	Iterable<T> listar();
 
-	T recuperar(int id);
+	T recuperar(Long id);
 
 	void crear(T objeto);
 
 	void modificar(T objeto);
 
-	void eliminar(int id);
+	void eliminar(Long id);
 }
